refactor: 작업자 관리 정보 조회 Facade 수정
- 정렬 기준을 담는 UserTaskSortType 추가

+++ src/main/java/com/fastcampus/befinal/application/facade/AdminFacade.java
@@ -35,7 +35,7 @@ public class AdminFacade {
     }
 
     public AdminDto.FindUserTaskListResponse findUserTaskScroll(AdminDto.FindUserTaskListRequest request) {
-        ScrollPagination<Long, AdminInfo.UserTaskInfo> info =
+        ScrollPagination<Integer, AdminInfo.UserTaskInfo> info =
             adminService.findUserTaskScroll(adminDtoMapper.toAdminCommand(request));
         return adminDtoMapper.fromUserTaskScroll(info);
     }
